requirements.txt
camelot-py
pandas
ghostscript
pypdf